# docs/swagger.yaml
basePath: /
definitions:
  model.JsonResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  model.VerifyContractRequest:
    properties:
      commit:
        type: string
      compilerVersion:
        type: string
      contractAddress:
        type: string
      contractDir:
        type: string
      contractUrl:
        type: string
      wasmFile:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soberkoder@gmail.com
    name: API Support
  description: This is a smart contract verify application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Verify Smart Contract API
  version: "1.0"
paths:
  /smart-contract/get-hash/{contractId}:
    get:
      consumes:
      - application/json
      description: Return the hash of a contract provided its code Id
      parameters:
      - description: Get contract hash
        in: path
        name: contractId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JsonResponse'
      summary: Get the hash of a deployed contract
      tags:
      - smart-contract
  /smart-contract/test-upload-s3:
    post:
      consumes:
      - application/json
      description: Upload S3
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JsonResponse'
      summary: Test upload S3
      tags:
      - smart-contract
  /smart-contract/verify:
    post:
      consumes:
      - application/json
      description: Compare if source code truely belongs to deployed smart contract
      parameters:
      - description: Verify smart contract source code
        in: body
        name: verify-contract-request
        required: true
        schema:
          $ref: '#/definitions/model.VerifyContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.JsonResponse'
      summary: Verify a smart contract source code
      tags:
      - smart-contract
swagger: "2.0"